# scan-coordinator/scan-coordinator-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SentinelFlow Security Platform",
    "description": "Unified security scan platform: trigger SAST + SCA scans, query vulnerabilities and BOM from Black Duck, and request AI-powered auto-fix via Claude Code.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://arborescent-actorly-carroll.ngrok-free.dev",
      "description": "SentinelFlow (ngrok)"
    }
  ],
  "paths": {
    "/scan/start": {
      "post": {
        "tags": [
          "Scan"
        ],
        "summary": "啟動 SAST + SCA 完整安全掃描",
        "description": "並行觸發 SAST（白箱）與 SCA（Black Duck 套件）兩種掃描。\n回傳 scan_id，用 /scan/status/{scan_id} 查詢進度。\n完成後用 /scan/report/{scan_id} 取得完整報告。",
        "operationId": "startScan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-ibm-nl-intent-examples": [
          "幫我掃描專案",
          "執行安全掃描",
          "掃描 sentinel-flow-demo",
          "開始做白箱和套件掃描",
          "run a full security scan",
          "scan the project for vulnerabilities",
          "start SAST and SCA scan",
          "trigger security scan"
        ]
      }
    },
    "/scan/status/{scan_id}": {
      "get": {
        "tags": [
          "Scan"
        ],
        "summary": "查詢掃描進度",
        "description": "step: triggering → scanning → normalizing → reporting → done\nstatus: running → completed",
        "operationId": "getScanStatus",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-ibm-nl-intent-examples": [
          "掃描進度如何",
          "掃描完成了嗎",
          "查詢掃描狀態",
          "check scan status",
          "is the scan done",
          "what is the scan progress"
        ]
      }
    },
    "/scan/report/{scan_id}": {
      "get": {
        "tags": [
          "Scan"
        ],
        "summary": "取得掃描報告",
        "description": "取得完整掃描報告（SAST + SCA 彙整）。掃描完成後才有內容。",
        "operationId": "getScanReport",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-ibm-nl-intent-examples": [
          "給我掃描報告",
          "顯示掃描結果",
          "有幾個漏洞",
          "安全掃描結果",
          "show security scan report",
          "get the vulnerability report",
          "what vulnerabilities were found"
        ]
      }
    },
...
    "/blackduck/scan/status/{job_id}": {
      "get": {
        "tags": [
          "Scan"
        ],
        "summary": "Get scan job status",
        "description": "Check the status of a triggered scan job.",
        "operationId": "getBDScanStatus",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-ibm-nl-intent-examples": [
          "check scan status",
          "is the scan done",
          "scan progress",
          "掃描狀態",
          "掃描完成了嗎"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanRequest": {
        "properties": {
          "project_name": {
            "title": "Project Name",
            "default": "sentinel-flow-demo",
            "type": "string"
          },
          "repo_path": {
            "title": "Repo Path",
            "default": "C:\\Users\\Administrator\\Desktop\\Projects\\sentinel-flow-demo",
            "type": "string"
          },
          "bd_project": {
            "title": "Bd Project",
            "default": "wyattlu-source/sentinel-flow-demo",
            "type": "string"
          },
          "bd_version": {
            "title": "Bd Version",
            "default": "main",
            "type": "string"
          },
          "scan_mode": {
            "title": "Scan Mode",
            "default": "full",
            "type": "string"
          },
          "trigger_reason": {
            "title": "Trigger Reason",
            "type": "string"
          },
          "project": {
            "title": "Project",
            "type": "string"
          }
        },
        "type": "object",
        "title": "ScanRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}